+++ tests/PointInTime/HourTest.php
@@ -14,6 +14,5 @@ class HourTest extends TestCase
 
         $this->assertSame(60, $hour->numberOfMinutes());
         $this->assertSame(13, $hour->toInt());
-        $this->assertSame('13', $hour->toString());
     }
 }